+++ src/Type/Definition/LeafType.php
@@ -42,8 +42,8 @@ interface LeafType
      *
      * Should throw an exception with a client friendly message on invalid value nodes, @see ClientAware.
      *
-     * @param array<string, mixed>|null $variables
      * @param ValueNode&Node $valueNode
+     * @param array<string, mixed>|null $variables
      *
      * @return mixed
      */