+++ src/Utils/Value.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace GraphQL\Utils;
 
+use GraphQL\Type\Definition\NamedType;
 use function array_key_exists;
 use function array_keys;
 use function array_map;
@@ -34,7 +35,7 @@ use Traversable;
  * @phpstan-type CoercedErrors array{errors: array<int, Error>, value: null}
  *
  * The key prev should actually also be typed as Path, but PHPStan does not support recursive types.
- * @phpstan-type Path array{prev: array<string, mixed>, key: string|int}
+ * @phpstan-type Path array{prev: array<string, mixed>|null, key: string|int}
  */
 class Value
 {
@@ -42,7 +43,7 @@ class Value
      * Given a type and any value, return a runtime value coerced to match the type.
      *
      * @param mixed $value
-     * @param ScalarType|EnumType|InputObjectType|ListOfType<Type&InputType>|NonNull $type
+     * @param InputType&Type $type
      * @phpstan-param Path|null $path
      *
      * @phpstan-return CoercedValue|CoercedErrors
@@ -60,6 +61,7 @@ class Value
                 ]);
             }
 
+            // @phpstan-ignore-next-line wrapped type is known to be input type after schema validation
             return self::coerceValue($value, $type->getWrappedType(), $blameNode, $path);
         }
 
@@ -116,6 +118,7 @@ class Value
         }
 
         if ($type instanceof ListOfType) {
+            /** @var Type&InputType $itemType known after schema validation */
             $itemType = $type->getWrappedType();
             if (is_array($value) || $value instanceof Traversable) {
                 $errors = [];
@@ -147,6 +150,7 @@ class Value
                 ? $coercedItem
                 : self::ofValue([$coercedItem['value']]);
         }
+        /** @var InputObjectType $type we handled all other cases at this point */
 
         if ($value instanceof stdClass) {
             // Cast objects to associative array before checking the fields.