+++ CHANGELOG.md
@@ -33,6 +33,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Make `Directive::$config['args']` use the same definition style as `FieldDefinition::$config['args']`
 - Rename `FieldArgument` to `Argument`
 - Make errors when parsing scalar literals more precise
+- Change expected `QueryPlan` options from `['group-implementor-fields']` to `['groupImplementorFields' => true]` in `ResolveInfo::lookAhead()`
 
 ### Added